backend/requirements.txt
langchain
langchain-community
openai
fastapi
uvicorn
python-dotenv
